feat :delete product on db

+++ src/favorite/favorite.service.ts
@@ -7,20 +7,32 @@ export class FavoritesService {
   constructor(private readonly prisma: PrismaService) {}
 
   async addFavorite(userId: number, productCode: number): Promise<Favorite> {
+    const numericProductCode = Number(productCode);
+    
+    if (isNaN(numericProductCode)) {
+      throw new Error('Invalid product code');
+    }
+
     return this.prisma.favorite.create({
       data: {
         userId,
-        productCode,
+        productCode: numericProductCode,
       },
     });
   }
 
+
   async removeFavorite(userId: number, productCode: number): Promise<Favorite> {
+    const numericProductCode = Number(productCode);
+    
+    if (isNaN(numericProductCode)) {
+      throw new Error('Invalid product code');
+    }
     return this.prisma.favorite.delete({
       where: {
         userId_productCode: {
           userId,
-          productCode,
+          productCode : numericProductCode,
         },
       },
     });
@@ -35,7 +47,7 @@ export class FavoritesService {
         product: true,
       },
     });
-    
+
     return favorites.map(favorite => favorite.product);
   }
 }
